git didnt take my changes

+++ src/MainMenu.cpp
@@ -92,25 +92,25 @@ void MainMenu::Unload(void)
 	Tiles::Unload();
 	Player::Unload();
 	AudioManager::unloadAsset();
-	int sz = Images.size();
-	for (int i = 0; i < sz; ++i) {
+	size_t sz = Images.size();
+	for (size_t i = 0; i < sz; ++i) {
 		Images[i].Free();
 		Images.pop_back();
 	}
 	sz = buttons.size();
-	for (int i = 0; i < sz; ++i) {
+	for (size_t i = 0; i < sz; ++i) {
 		buttons.pop_back();
 	}
 	sz = enemy.size();
-	for (int i = 0; i < sz; ++i) {
+	for (size_t i = 0; i < sz; ++i) {
 		enemy.pop_back();
 	}
 	sz = player.size();
-	for (int i = 0; i < sz; ++i) {
+	for (size_t i = 0; i < sz; ++i) {
 		player.pop_back();
 	}
 	sz = tiles.size();
-	for (int i = 0; i < sz; ++i) {
+	for (size_t i = 0; i < sz; ++i) {
 		tiles.pop_back();
 	}
 }
